chore: move articles and alerts management
