+++ web-client/src/pages/Admin/AlertsManagement/components/AlertSMSForm.tsx
@@ -7,7 +7,7 @@ import {
 } from "react-hook-form";
 
 // Redux
-import { useSendAlertMutation } from "../../../services/alertQuery";
+import { useSendAlertMutation } from "../../../../services/alertQuery";
 
 import { useState } from "react";
 import Select from "react-select";
@@ -15,7 +15,7 @@ import Select from "react-select";
 import {
   getBarangays,
   Barangay,
-} from "../../../components/AddressSelector/AddressSelector";
+} from "../../../../components/AddressSelector/AddressSelector";
 
 const AlertForm = () => {
   const [formMessage, setFormMessage] = useState<string>("");
